tty_hci: handle reads of size 0 and nonblocking
bluedroid seems to use reads of size 0 to wait for new data.
Returning an error here confuses bluedroid and leads to discarded
HCI packets, as the stack then tries to read a packet with size
1 which is smaller than the next received packet. In response
to that, tty_hci.c throws this packet away.

Also fix implementation of non-/blocking read().

Change-Id: I7b199f939a8230ebba2416a2b28d1c3cf5af759d